feat: set up ts
