+++ package.json
@@ -6,6 +6,10 @@
     "@testing-library/jest-dom": "^5.17.0",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
+    "@types/jest": "^29.5.12",
+    "@types/node": "^20.14.10",
+    "@types/react": "^18.3.3",
+    "@types/react-dom": "^18.3.0",
     "docx": "^8.5.0",
     "file-saver": "^2.0.5",
     "firebase": "^10.12.2",
@@ -14,6 +18,7 @@
     "react-dom": "^18.3.1",
     "react-quill": "^2.0.0",
     "react-scripts": "5.0.1",
+    "typescript": "^5.5.3",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
